new changes added in 16/06/2017 5:50pm

+++ includes/js/customer.js
@@ -64,6 +64,25 @@ app.controller('userctrl', function($scope, $http) {
             console.log(error);
         });
     };
+
+    $scope.get_country = function() {
+        var params = {};
+        params.action = "inventory_crud_function";
+        params.type = "get_country";
+        $http({
+            url: myAjax.ajaxurl,
+            method: "POST",
+            params: params
+        }).then(function(response) {
+            console.log(response.data);
+            $scope.country = response.data;
+            // if(response.data){
+            //    console.log('new user adding successful');
+            // }
+        }, function(error) {
+            console.log(error);
+        });
+    };
     $scope.get_all_user = function() {
         $scope.loading = true;
         var params = {};
@@ -86,31 +105,36 @@ app.controller('userctrl', function($scope, $http) {
         });
     };
     $scope.delete = function(id) {
-        $scope.loading = true;
-        console.log(id);
-        var params = {};
-        params.action = "inventory_crud_function";
-        params.type = "delete";
-        params.table = "inv_customer";
-        params.id = id;
-        $http({
-            url: myAjax.ajaxurl,
-            method: "POST",
-            params: params
-        }).then(function(response) {
-            console.log(response.data);
-            // $scope.delete = response.data;
-            if (response.data === '1') {
-                console.log('successful');
-                $scope.get_all_user();
-            } else if (response.data === '23000') {
-                alert("you can not delete this.Because it is used in somewhere else.");
-                $scope.loading = false;
-            }
-        }, function(error) {
-            console.log(error);
-        });
+        var DeleteConfirmation = confirm("Do you wish to proceed?");
+        if (DeleteConfirmation == true) {
+            $scope.loading = true;
+            console.log(id);
+            var params = {};
+            params.action = "inventory_crud_function";
+            params.type = "delete";
+            params.table = "inv_customer";
+            params.id = id;
+            $http({
+                url: myAjax.ajaxurl,
+                method: "POST",
+                params: params
+            }).then(function(response) {
+                console.log(response.data);
+                // $scope.delete = response.data;
+                if (response.data === '1') {
+                    console.log('successful');
+                    $scope.get_all_user();
+                } else if (response.data === '23000') {
+                    alert("you can not delete this.Because it is used in somewhere else.");
+                    $scope.loading = false;
+                }
+            }, function(error) {
+                console.log(error);
+            });
+        }
+
     }
+    $scope.get_country();
     $scope.get_currency();
     $scope.get_all_user();
 });
