+++ includes/js/location.js
@@ -72,33 +72,56 @@ app.controller('locationctrl', function($scope, $http) {
         });
     };
     $scope.delete = function(id) {
-        $scope.loading = true;
-        console.log(id);
+        var DeleteConfirmation = confirm("Do you wish to proceed?");
+        if (DeleteConfirmation == true) {
+            $scope.loading = true;
+            console.log(id);
+            var params = {};
+            params.action = "inventory_crud_function";
+            params.type = "delete";
+            params.table = "inv_location";
+            params.id = id;
+            $http({
+                url: myAjax.ajaxurl,
+                method: "POST",
+                params: params
+            }).then(function(response) {
+                console.log(response.data);
+                // $scope.delete = response.data;
+                if (response.data === '1') {
+                    console.log('successful');
+                    // setTimeout(function () {
+                    $scope.get_location();
+                    // },1000);
+
+                } else if (response.data === '23000') {
+                    alert("you can not delete this.Because it is used in somewhere else.");
+                    $scope.loading = false;
+                }
+            }, function(error) {
+                console.log(error);
+            });
+        }
+
+    }
+    $scope.get_country = function() {
         var params = {};
         params.action = "inventory_crud_function";
-        params.type = "delete";
-        params.table = "inv_location";
-        params.id = id;
+        params.type = "get_country";
         $http({
             url: myAjax.ajaxurl,
             method: "POST",
             params: params
         }).then(function(response) {
             console.log(response.data);
-            // $scope.delete = response.data;
-            if (response.data === '1') {
-                console.log('successful');
-                // setTimeout(function () {
-                $scope.get_location();
-                // },1000);
-
-            } else if (response.data === '23000') {
-                alert("you can not delete this.Because it is used in somewhere else.");
-                $scope.loading = false;
-            }
+            $scope.country = response.data;
+            // if(response.data){
+            //    console.log('new user adding successful');
+            // }
         }, function(error) {
             console.log(error);
         });
-    }
+    };
+    $scope.get_country();
     $scope.get_location();
 });
