+++ includes/js/product.js
@@ -1,4 +1,5 @@
-app.controller('productctrl', function($scope, $http) {
+app.controller('productctrl', function($scope, $http, $timeout) {
+
     $scope.add = function(cat) {
         // console.log(cat);
         cat.action = "inventory_crud_function";
@@ -36,6 +37,11 @@ app.controller('productctrl', function($scope, $http) {
                 $scope.products = [];
             } else {
                 $scope.products = response.data;
+                $scope.totalItems = $scope.products.length;
+                $scope.currentPage = 1;
+                $scope.itemsPerPage = 10;
+                $scope.maxSize = 15; //Number of pager buttons to show
+                $scope.viewby = 10;
             }
             $scope.loading = false;
             // if(response.data){
@@ -46,33 +52,36 @@ app.controller('productctrl', function($scope, $http) {
         });
     };
     $scope.delete = function(id) {
-        $scope.loading = true;
-        console.log(id);
-        var params = {};
-        params.action = "inventory_crud_function";
-        params.type = "delete";
-        params.table = "inv_product";
-        params.id = id;
-        $http({
-            url: myAjax.ajaxurl,
-            method: "POST",
-            params: params
-        }).then(function(response) {
-            console.log(response.data);
-            // $scope.delete = response.data;
-            if (response.data === '1') {
-                console.log('successful');
-                // setTimeout(function () {
-                $scope.get_product();
-                // },1000);
+        var DeleteConfirmation = confirm("Do you wish to proceed?");
+        if (DeleteConfirmation == true) {
+            $scope.loading = true;
+            console.log(id);
+            var params = {};
+            params.action = "inventory_crud_function";
+            params.type = "delete";
+            params.table = "inv_product";
+            params.id = id;
+            $http({
+                url: myAjax.ajaxurl,
+                method: "POST",
+                params: params
+            }).then(function(response) {
+                console.log(response.data);
+                // $scope.delete = response.data;
+                if (response.data === '1') {
+                    console.log('successful');
+                    // setTimeout(function () {
+                    $scope.get_product();
+                    // },1000);
 
-            } else if (response.data === '23000') {
-                alert("you can not delete this.Because it is used in somewhere else.");
-                $scope.loading = false;
-            }
-        }, function(error) {
-            console.log(error);
-        });
+                } else if (response.data === '23000') {
+                    alert("you can not delete this.Because it is used in somewhere else.");
+                    $scope.loading = false;
+                }
+            }, function(error) {
+                console.log(error);
+            });
+        }
     }
     $scope.edit_modal = function(data) {
         console.log(data);
@@ -134,27 +143,50 @@ app.controller('productctrl', function($scope, $http) {
             $scope.obj = response.data;
             $scope.parent = [];
             $scope.children = [];
+            $scope.grandParent = [];
             for (var i = 0; i < $scope.obj.length; i++) {
                 if ($scope.obj[i]['inv_product_cat_parent'] == 0) {
-                    $scope.parent.push($scope.obj[i]);
+                    if ($scope.parent.indexOf($scope.obj[i]['id']) == -1) {
+                        $scope.parent.push($scope.obj[i]['id']);
+                    }
                 }
                 if ($scope.obj[i]['inv_product_cat_parent'] != 0) {
+                    if ($scope.parent.indexOf($scope.obj[i]['inv_product_cat_parent']) == -1) {
+                        $scope.parent.push($scope.obj[i]['inv_product_cat_parent']);
+                    }
                     $scope.children.push($scope.obj[i]);
                 }
             }
 
-            $scope.grandParent = [];
-            for (var i = 0; i < $scope.parent.length; i++) {
-                $scope.temparr = [];
-                $scope.temparr.children = [];
-                for (var j = 0; j < $scope.children.length; j++) {
-                    if ($scope.parent[i]['id'] == $scope.children[j]['inv_product_cat_parent']) {
-                        $scope.temparr.children.push($scope.children[j]);
+            for (var i = 0; i < $scope.obj.length; i++) {
+                if ($scope.parent.indexOf($scope.obj[i]['id']) !== -1) {
+                    console.log($scope.obj[i]['id'] + "id there");
+                    $scope.grandParent.push($scope.obj[i]);
+                }
+            }
+            for (var i = 0; i < $scope.grandParent.length; i++) {
+                $scope.grandParent[i].children=[];
+                debugger;
+                for (var j = 0; j < $scope.obj.length; j++) {
+                    if($scope.grandParent[i]['id'] == $scope.obj[j]['inv_product_cat_parent']){
+                       $scope.grandParent[i].children.push($scope.obj[j]); 
                     }
                 }
-                $scope.temparr.push($scope.parent[i]);
-                $scope.grandParent.push($scope.temparr);
             }
+
+
+            // $scope.grandParent = [];
+            // for (var i = 0; i < $scope.parent.length; i++) {
+            //     $scope.temparr = [];
+            //     $scope.temparr.children = [];
+            //     for (var j = 0; j < $scope.children.length; j++) {
+            //         if ($scope.parent[i]['id'] == $scope.children[j]['inv_product_cat_parent']) {
+            //             $scope.temparr.children.push($scope.children[j]);
+            //         }
+            //     }
+            //     $scope.temparr.push($scope.parent[i]);
+            //     $scope.grandParent.push($scope.temparr);
+            // }
             console.log("parent", $scope.parent);
             console.log("children", $scope.children);
             console.log('grandParent', $scope.grandParent);
@@ -183,6 +215,18 @@ app.controller('productctrl', function($scope, $http) {
     };
     $scope.get_supplier();
     $scope.get_category();
-    $scope.get_parent_category();
+    // $scope.get_parent_category();
     $scope.get_product();
+
+    /*----------  Pagination config area  ----------*/
+    // $timeout(function() {
+
+    // }, 2000);
+    $scope.pageChanged = function() {
+        console.log('Page changed to: ' + $scope.currentPage);
+    };
+    $scope.setItemsPerPage = function(num) {
+        $scope.itemsPerPage = num;
+        $scope.currentPage = 1; //reset to first page
+    }
 });
